requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
python-dotenv
requests
beautifulsoup4
